Move graphiql specs support methods to a reusable module
This will help when cutomizing the GrapQL API
