+++ spec/spec_helper.rb
@@ -20,6 +20,7 @@ RSpec.configure do |config|
   config.include(EmailSpec::Matchers)
   config.include(CommonActions)
   config.include(ActiveSupport::Testing::TimeHelpers)
+  config.include Graphql, type: :graphql_api
 
   config.define_derived_metadata(file_path: Regexp.new("/spec/components/")) do |metadata|
     metadata[:type] = :component
